# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API Documentation
  description: Документация API для всех операций, включая аутентификацию, управление картами, блокировку карт и управление пользователями
  version: 1.0.0
servers:
  - url: http://localhost:8080
    description: Локальный сервер для разработки
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    UserResponseDto:
      type: object
      description: Объект ответа, содержащий информацию о пользователе
      required:
        - id
        - login
        - roleName
        - isEnabled
        - isNonLocked
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор пользователя
          example: 10
        login:
          type: string
          description: Логин пользователя
          example: user123
        roleName:
          type: string
          description: Название роли пользователя
          example: ADMIN
        isEnabled:
          type: boolean
          description: Флаг, указывающий, активен ли пользователь
          example: true
        isNonLocked:
          type: boolean
          description: Флаг, указывающий, не заблокирован ли пользователь
          example: true
        cards:
          type: array
          description: Список карт пользователя (может отсутствовать в некоторых ответах)
          items:
            $ref: '#/components/schemas/CardResponseDto'
...
    CardResponseDto:
      type: object
      description: Объект ответа, содержащий информацию о карте
      required:
        - id
        - maskedCardNumber
        - ownerLogin
        - expiryDate
        - statusName
        - balance
      properties:
        id:
          type: integer
          format: int64
          description: Уникальный идентификатор карты
          example: 10
        maskedCardNumber:
          type: string
          description: Маскированный номер карты, отображающий только последние 4 цифры
          example: "************1234"
        ownerLogin:
          type: string
          description: Логин владельца карты
          example: user123
        expiryDate:
          type: string
          pattern: ^\d{4}-\d{2}$
          description: Дата истечения срока действия карты в формате yyyy-MM
          example: 2025-08
        statusName:
          type: string
          description: Название статуса карты
          example: ACTIVE
        balance:
          type: number
          format: double
          description: Текущий баланс на карте
          example: 1500.75
...
    UpdateUserRoleDto:
      type: object
      description: Объект запроса для обновления роли пользователя
      required:
        - role
      properties:
        role:
          $ref: '#/components/schemas/UserRoleCode'
    UserRoleCode:
      type: string
      description: Роль пользователя
      enum:
        - USER
        - ADMIN
...
paths:
...
  /users/{id}:
    get:
      summary: Получение пользователя по ID
      description: Возвращает информацию о пользователе по его идентификатору
      tags:
        - Управление пользователями
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор пользователя
          example: 10
      responses:
        '200':
          description: Успешное получение пользователя
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Пользователь не найден
          content: {}
    delete:
      summary: Удаление пользователя
      description: Удаляет пользователя по его идентификатору
      tags:
        - Управление пользователями
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор пользователя
          example: 10
      responses:
        '204':
          description: Пользователь успешно удален
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Пользователь не найден
          content: {}
...
  /users/{id}/role:
    patch:
      summary: Обновление роли пользователя
      description: Обновляет роль пользователя по его идентификатору
      tags:
        - Управление пользователями
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор пользователя
          example: 10
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRoleDto'
      responses:
        '200':
          description: Роль успешно обновлена
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '400':
          description: Некорректные данные запроса
          content: {}
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Пользователь не найден
          content: {}
  /users/{id}/toggleLock:
    patch:
      summary: Переключение статуса блокировки пользователя
      description: Переключает статус блокировки пользователя (заблокирован/не заблокирован)
      tags:
        - Управление пользователями
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор пользователя
          example: 10
      responses:
        '200':
          description: Статус блокировки успешно изменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Пользователь не найден
          content: {}
  /users/{id}/toggleEnable:
    patch:
      summary: Переключение статуса активности пользователя
      description: Переключает статус активности пользователя (активен/не активен)
      tags:
        - Управление пользователями
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
          description: Идентификатор пользователя
          example: 10
      responses:
        '200':
          description: Статус активности успешно изменен
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponseDto'
        '403':
          description: Доступ запрещен
          content: {}
        '404':
          description: Пользователь не найден
          content: {}